+++ packages/react-native-layout/.eslintrc.js
@@ -1,0 +1,15 @@
+module.exports = { 
+  extends: [
+    '../../.eslintrc.js',
+    '@bothrs/eslint-config/react',
+    '@bothrs/eslint-config/react-native',
+  ],
+  parserOptions: {
+    project: 'tsconfig.json',
+    tsconfigRootDir: __dirname,
+    sourceType: 'module'
+  },
+  rules: {
+    "unused-imports/no-unused-imports": "off",
+  }
+}