+++ packages/react-native-layout/src/index.ts
@@ -1,5 +1,7 @@
 export * from './components/container'
+export * from './components/flex'
 export * from './components/line'
+export * from './components/scroll-view-sticky-bottom'
 export * from './components/spacing'
 
 export * from './types/generic-props'