+++ packages/translations/src/index.ts
@@ -1,59 +1,89 @@
-import { useState, useEffect } from 'react'
 import i from 'i18next'
 import ChainedBackend from 'i18next-chained-backend'
-
-//@ts-ignore i18next-async-storage-backend does not have type declarations
-import AsyncStorageBackend from 'i18next-async-storage-backend2'
-import MultiloadAdapter from 'i18next-multiload-backend-adapter'
 import Fetch from 'i18next-fetch-backend'
+import MultiloadAdapter from 'i18next-multiload-backend-adapter'
+import { useState } from 'react'
+
+import StorageBackend from './storage-backend'
 
 import type {
+  AirtableInit,
   FormattedTranslation,
+  HealthblocksInit,
   Translation,
-  TranslationInitParams,
+  TranslationInitParameters,
   TranslationKeys,
+  TranslationRow,
 } from './types'
+
+let initializing = false
 
 /**
  * A hook which initializes i18next and loads the languages from an external data source
- * @param initParams takes the loadPath, startupLanguage and expiration time
+ * @param initParams takes fetch options cache expiration time and all i18next params
  * @returns boolean which indicates if the languages are loaded
  */
-export function useTranslations(initParams: TranslationInitParams): boolean {
-  const [initialized, setInitialized] = useState(i.isInitialized)
-
-  const handleInitialized = () => {
-    setInitialized(true)
+export function useTranslations(
+  initParameters: TranslationInitParameters
+): boolean {
+  if (!initializing) {
+    initializing = true
+    initTranslations(initParameters).then(() => setInitialized(true))
   }
-
-  !initialized && initTranslations(initParams)
-
-  useEffect(() => {
-    i.on('initialized', handleInitialized)
-    return () => i.off('initialized', handleInitialized)
-  }, [])
+  const [initialized, setInitialized] = useState(i.isInitialized || false)
 
   return initialized
 }
 
-function initTranslations({
-  loadPath,
-  startupLanguage,
+export function initTranslations({
   expirationTime,
-  fallbackLng,
-  dataFormatter,
-}: TranslationInitParams) {
-  const fetchOptions = {
+  fetchOptions,
+  ...options
+}: TranslationInitParameters) {
+  return i.use(ChainedBackend).init({
+    ns: 'app',
+    defaultNS: 'app',
+    fallbackLng: options.lng || 'en',
+    load: 'all',
+    keySeparator: false,
+    nsSeparator: false,
+    interpolation: {
+      prefix: '[',
+      suffix: ']',
+    },
+    backend: {
+      backends: [StorageBackend, MultiloadAdapter],
+      backendOptions: [
+        { prefix: 'i18n_', expirationTime },
+        { backend: Fetch, backendOption: fetchOptions },
+      ],
+    },
+    ...options,
+  })
+}
+
+/**
+ * A hook which initializes i18next and loads the languages from Airtable
+ * @param initParams takes the airtable load path, cache expiration time and all i18next params
+ * @returns boolean which indicates if the languages are loaded
+ */
+export function useAirtableTranslations({ loadPath, ...init }: AirtableInit) {
+  return useTranslations({
+    ...init,
+    fetchOptions: airtableFetchOptions(loadPath),
+  })
+}
+
+export function airtableFetchOptions(loadPath: string) {
+  return {
     loadPath,
     allowMultiLoading: true,
-    parse: function (data: string) {
+    parse(data: string) {
       const parsedData = JSON.parse(data)
 
-      const translationData: FormattedTranslation[] = dataFormatter
-        ? dataFormatter(parsedData as object)
-        : parsedData
+      const translationData: FormattedTranslation[] = parsedData
 
-      const langSet = new Set()
+      const langSet = new Set<string>()
 
       translationData.forEach((translation) => {
         Object.keys(translation)
@@ -64,7 +94,7 @@ function initTranslations({
       })
 
       const keys: TranslationKeys = {}
-      Array.from(langSet).forEach((lang) => {
+      langSet.forEach((lang) => {
         keys[lang as string] = { app: {} }
       })
 
@@ -77,32 +107,45 @@ function initTranslations({
       return keys
     },
   }
+}
 
-  const asyncOptions = {
-    prefix: 'i18n_',
-    expirationTime,
-  }
-
-  let backend = {
-    backends: [AsyncStorageBackend, MultiloadAdapter],
-    backendOptions: [
-      asyncOptions,
-      { backend: Fetch, backendOption: fetchOptions },
-    ],
-  }
-
-  i.use(ChainedBackend).init({
-    ns: 'app',
-    defaultNS: 'app',
-    fallbackLng: fallbackLng || 'en',
-    lng: startupLanguage,
-    load: 'all',
-    keySeparator: false,
-    nsSeparator: false,
-    interpolation: {
-      prefix: '[',
-      suffix: ']',
-    },
-    backend,
+/**
+ * A hook which initializes i18next and loads the languages from an external data source
+ * @param initParams takes the project id, api url, cache expiration time and all i18next params
+ * @returns boolean which indicates if the languages are loaded
+ */
+export function useHealthblocksTranslations({
+  api,
+  projectId,
+  ...init
+}: HealthblocksInit) {
+  return useTranslations({
+    ...init,
+    fetchOptions: healthblocksFetchOptions(projectId, api),
   })
+}
+
+export function healthblocksFetchOptions(
+  projectId: number,
+  api: string = 'https://v2.healthblocks.io'
+) {
+  return {
+    loadPath: api + '/translation?projectId=' + projectId,
+    parse(json: string) {
+      const resources: TranslationKeys = {}
+
+      const rows: TranslationRow[] = JSON.parse(json)
+      for (const row of rows) {
+        for (const lang in row.values) {
+          // Add language if missing
+          if (!resources[lang]) resources[lang] = { app: {} }
+
+          // Add key:value
+          resources[lang].app[row.key] = row.values[lang]
+        }
+      }
+
+      return resources
+    },
+  }
 }