+++ packages/translations/src/storage-backend.native.ts
@@ -1,0 +1,1 @@
+export { default } from 'i18next-async-storage-backend2'