+++ packages/airtable/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.7](https://github.com/bothrs/open-source/compare/@bothrs/airtable@0.3.6...@bothrs/airtable@0.3.7) (2022-07-19)
+
+
+### Bug Fixes
+
+* added yarn workspaces again ([3c78cb9](https://github.com/bothrs/open-source/commit/3c78cb92254d20dbc231336cc7afe54300c1efb5))
+
+
+
+
+
+## [0.3.6](https://github.com/bothrs/open-source/compare/@bothrs/airtable@0.3.5...@bothrs/airtable@0.3.6) (2022-06-07)
+
+**Note:** Version bump only for package @bothrs/airtable
+
+
+
+
+
+## [0.3.5](https://github.com/bothrs/open-source/compare/@bothrs/airtable@0.3.4...@bothrs/airtable@0.3.5) (2022-03-14)
+
+
+### Bug Fixes
+
+* use tsc for building @bothrs/translations ([40ad9ab](https://github.com/bothrs/open-source/commit/40ad9abd0c6bfca081b26705462786299abbc9fb))
+
+
+
+
+
 ## [0.3.4](https://github.com/bothrs/open-source/compare/@bothrs/airtable@0.3.3...@bothrs/airtable@0.3.4) (2021-11-18)