+++ packages/zero-height/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.10](https://github.com/bothrs/open-source/compare/@bothrs/zero-height@0.1.9...@bothrs/zero-height@0.1.10) (2022-07-19)
+
+
+### Bug Fixes
+
+* added yarn workspaces again ([3c78cb9](https://github.com/bothrs/open-source/commit/3c78cb92254d20dbc231336cc7afe54300c1efb5))
+
+
+
+
+
+## [0.1.9](https://github.com/bothrs/open-source/compare/@bothrs/zero-height@0.1.8...@bothrs/zero-height@0.1.9) (2022-06-07)
+
+**Note:** Version bump only for package @bothrs/zero-height
+
+
+
+
+
+## [0.1.8](https://github.com/bothrs/open-source/compare/@bothrs/zero-height@0.1.7...@bothrs/zero-height@0.1.8) (2022-03-24)
+
+
+### Bug Fixes
+
+* check the amount of font families ([ec5435b](https://github.com/bothrs/open-source/commit/ec5435b678daac1b32093cb3885186b247eea6a0))
+
+
+
+
+
 ## [0.1.7](https://github.com/bothrs/open-source/compare/@bothrs/zero-height@0.1.6...@bothrs/zero-height@0.1.7) (2021-11-18)
 
 **Note:** Version bump only for package @bothrs/zero-height